+++ src/dipdup/yaml.py
@@ -85,6 +85,17 @@ def fix_dataclass_field_aliases(config: dict[str, Any]) -> None:
                     fix_dataclass_field_aliases(item)
 
 
+def lowercase_logging_level(config: dict[str, Any]) -> None:
+    logging_ = config.get('logging')
+
+    if not logging_:
+        return
+    elif isinstance(logging_, str):
+        config['logging'] = logging_.lower()
+    elif isinstance(logging_, dict):
+        for k in logging_:
+            logging_[k] = logging_[k].lower()
+
 class DipDupYAMLConfig(dict[str, Any]):
     @classmethod
     def load(
@@ -106,11 +117,7 @@ class DipDupYAMLConfig(dict[str, Any]):
 
             config.update(yaml.load(path_yaml))
 
-        config.validate_version()
-
-        # FIXME: Can't use `from_` field alias in dataclass; fixed in dipdup.yaml instead
-        # FIXME: See https://github.com/pydantic/pydantic/issues/4286
-        fix_dataclass_field_aliases(config)
+        config._post_load_hooks()
 
         return config, config_environment
 
@@ -131,3 +138,10 @@ class DipDupYAMLConfig(dict[str, Any]):
             raise ConfigurationError(
                 f'Incompatible spec version: expected {__spec_version__}, got {config_spec_version}. See https://docs.dipdup.io/config/spec_version'
             )
+
+    def _post_load_hooks(self) -> None:
+        self.validate_version()
+        # FIXME: Can't use `from_` field alias in dataclass; fixed in dipdup.yaml instead
+        # FIXME: See https://github.com/pydantic/pydantic/issues/4286
+        fix_dataclass_field_aliases(self)
+        lowercase_logging_level(self)  # case insensitive logging